chore: ini query suspense
